requirements.txt
eventlet==0.33.3
Flask==2.3.2
Flask-SocketIO==5.3.4
numpy==1.24.4
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
